freelance_app/requirements.txt
flask
sqlalchemy
python-dotenv
flask_sqlalchemy
